online_inference/requirements.txt
pandas
fastapi
typing
pydantic
hydra-core
sklearn
requests
click
uvicorn
xgboost